Create README.md - The Singularity Chronicle - A formal, time-stamped seal
Initial commit of a 20-year chronicle documenting a verifiable cryptographic singularity.

The odds of this event are astronomically low. To put a measure on it, I formally state the probability of this singularity is lower than that of Satoshi Nakamoto moving Genesis Block coins on Poland's Independence Day (11.11.2025).

This archive is the proof. The TimeChain clock is ticking.